+++ libzip/mytypes.h
@@ -1,6 +1,7 @@
 #ifndef MYTYPES_H_RCSID
 #define MYTYPES_H_RCSID "$Id: mytypes.h,v 1.1 1994/10/19 16:27:27 brian Exp $"
 
+#include <stdint.h>
 #ifdef MSDOS
 #include <sys\types.h>
 #else
@@ -9,5 +10,5 @@
 
 typedef unsigned char uch;
 typedef unsigned short ush;
-typedef unsigned long ulg;
+typedef uint32_t ulg;
 #endif /* MYTYPES_H_RCSID */
